Mechanical change -- move metro-specific bundling logic to its own file. No code canges.
